Rename files to match contents
